# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.User:
    properties:
      _id:
        type: string
      app_notification:
        type: boolean
      email:
        type: string
      fcm_token:
        type: string
      image:
        type: string
      is_oauth:
        type: boolean
      is_premium:
        type: boolean
      mail_notification:
        type: boolean
      membership_type:
        description: 0 Basic, 1 Premium 2 Premium Supporter
        type: integer
      oauth_type:
        type: integer
      username:
        type: string
    type: object
  requests.ChangeFCMToken:
    properties:
      fcm_token:
        type: string
    required:
    - fcm_token
    type: object
  requests.ChangeImage:
    properties:
      image:
        type: string
    required:
    - image
    type: object
  requests.ChangeMembership:
    properties:
      is_premium:
        type: boolean
      membership_type:
        type: integer
    type: object
  requests.ChangeNotification:
    properties:
      app_notification:
        type: boolean
      mail_notification:
        type: boolean
    required:
    - app_notification
    - mail_notification
    type: object
  requests.ChangePassword:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  requests.ForgotPassword:
    properties:
      email_address:
        type: string
    required:
    - email_address
    type: object
  requests.GetProfile:
    properties:
      username:
        type: string
    required:
    - username
    type: object
...
  requests.Register:
    properties:
      email_address:
        type: string
      fcm_token:
        type: string
      image:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email_address
    - password
    - username
    type: object
...
  responses.Author:
    properties:
      _id:
        type: string
      email:
        type: string
      image:
        type: string
      is_premium:
        type: boolean
      username:
        type: string
    type: object
  responses.ConsumeLater:
    properties:
      _id:
        type: string
      content:
        $ref: '#/definitions/responses.ConsumeLaterContent'
      content_external_id:
        type: string
      content_external_int_id:
        type: integer
      content_id:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      self_note:
        type: string
      user_id:
        type: string
    type: object
  responses.ConsumeLaterContent:
    properties:
      description:
        type: string
      image_url:
        type: string
      score:
        type: number
      title_en:
        type: string
      title_original:
        type: string
    type: object
...
  responses.ReviewContent:
    properties:
      image_url:
        type: string
      title_en:
        type: string
      title_original:
        type: string
    type: object
...
  responses.ReviewWithContent:
    properties:
      _id:
        type: string
      author:
        $ref: '#/definitions/responses.Author'
      content:
        $ref: '#/definitions/responses.ReviewContent'
      content_external_id:
        type: string
      content_external_int_id:
        type: integer
      content_id:
        type: string
      created_at:
        type: string
      is_author:
        type: boolean
      is_liked:
        type: boolean
      likes:
        items:
          type: string
        type: array
      popularity:
        type: integer
      review:
        type: string
      star:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  responses.UserInfo:
    properties:
      _id:
        type: string
      anime_count:
        type: integer
      anime_watched_episodes:
        type: integer
      consume_later:
        items:
          $ref: '#/definitions/responses.ConsumeLater'
        type: array
      email:
        type: string
      fcm_token:
        type: string
      game_count:
        type: integer
      game_total_hours_played:
        type: integer
      image:
        type: string
      is_premium:
        type: boolean
      legend_content:
        items:
          $ref: '#/definitions/responses.UserInfoContent'
        type: array
      level:
        type: integer
      membership_type:
        type: integer
      movie_count:
        type: integer
      movie_watched_time:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/responses.ReviewWithContent'
        type: array
      tv_count:
        type: integer
      tv_watched_episodes:
        type: integer
      username:
        type: string
    type: object
  responses.UserInfoContent:
    properties:
      _id:
        type: string
      content_type:
        type: string
      hours_played:
        type: integer
      image_url:
        type: string
      times_finished:
        type: integer
      title_en:
        type: string
      title_original:
        type: string
    type: object
...
host: http://localhost:8080
info:
  contact:
    email: mrntlu@gmail.com
    name: Burak Fidan
  description: REST Api of Project Consumer.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: '#soon'
  title: Project Consumer API
  version: "1.0"
paths:
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Allows users to register
      parameters:
      - description: User registration info
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/requests.Register'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: User Registration
      tags:
      - auth
...
  /user:
    delete:
      consumes:
      - application/json
      description: Deletes everything related to user
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Deletes user information
      tags:
      - user
  /user/basic:
    get:
      consumes:
      - application/json
      description: Returns basic user information
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User
          schema:
            $ref: '#/definitions/models.User'
      security:
      - ApiKeyAuth: []
      summary: User basic info
      tags:
      - user
  /user/forgot-password:
    post:
      consumes:
      - application/json
      description: Users can change their password when they forgot
      parameters:
      - description: User's email
        in: body
        name: ForgotPassword
        required: true
        schema:
          $ref: '#/definitions/requests.ForgotPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Couldn't find any user
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Will be used when user forgot password
      tags:
      - user
  /user/image:
    patch:
      consumes:
      - application/json
      description: User can update their image
      parameters:
      - description: Change Image
        in: body
        name: changeimage
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeImage'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates user image
      tags:
      - user
  /user/info:
    get:
      consumes:
      - application/json
      description: Returns users stats from username
      parameters:
      - description: Get Profile
        in: body
        name: getprofile
        required: true
        schema:
          $ref: '#/definitions/requests.GetProfile'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Info
          schema:
            $ref: '#/definitions/responses.UserInfo'
      security:
      - ApiKeyAuth: []
      summary: User info from username
      tags:
      - user
  /user/membership:
    patch:
      consumes:
      - application/json
      description: User membership status will be updated depending on subscription
        status
      parameters:
      - description: Set Membership
        in: body
        name: changemembership
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeMembership'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Change User Membership
      tags:
      - user
  /user/notification:
    patch:
      consumes:
      - application/json
      description: Users can change their notification preference
      parameters:
      - description: Set notification
        in: body
        name: changenotification
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeNotification'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Change User Notification Preference
      tags:
      - user
  /user/password:
    patch:
      consumes:
      - application/json
      description: Users can change their password
      parameters:
      - description: Set new password
        in: body
        name: ChangePassword
        required: true
        schema:
          $ref: '#/definitions/requests.ChangePassword'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Change User Password
      tags:
      - user
  /user/token:
    patch:
      consumes:
      - application/json
      description: Depending on logged in device fcm token will be updated
      parameters:
      - description: Set token
        in: body
        name: changefcmtoken
        required: true
        schema:
          $ref: '#/definitions/requests.ChangeFCMToken'
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Updates FCM User Token
      tags:
      - user
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"